+++ web_src/src/pages/app/index.tsx
@@ -110,6 +110,7 @@ import { useCanvasEchoReleaseGuards } from "./useCanvasEchoReleaseGuards";
 import { useCanvasLifecycleEventHandlers } from "./useCanvasLifecycleEventHandlers";
 import { useDraftStagingActions } from "./useDraftStagingActions";
 import { executeCommitStaging } from "./lib/commit-staging-flow";
+import { buildDuplicatedEdges, buildDuplicatedNodes } from "./lib/duplicate-nodes";
 import { getNodeIntegrationName, overlayIntegrationWarnings } from "./lib/node-integrations";
 import { renderCanvasNodeCustomField } from "./lib/render-canvas-node-custom-field";
 import { buildCanvasYamlExportPayload, materializeCanvasSpec } from "./lib/workflow-spec-files";
@@ -168,68 +169,6 @@ const CANVAS_AUTO_LAYOUT_ON_UPDATE_STORAGE_KEY = "canvas-auto-layout-on-update-e
 const VERSION_ACTION_SAVE_SETTLE_TIMEOUT_MS = 5000;
 const EMPTY_CANVAS_SPEC_ITEMS: never[] = [];
 const RUNNING_RUNS_FILTERS = { states: ["STATE_STARTED" as CanvasesCanvasRunState] };
-
-type DuplicatedNodesResult = {
-  newNodes: ComponentsNode[];
-  nodeIdMap: Record<string, string>;
-};
-
-function duplicateBaseName(node: ComponentsNode): string {
-  const trimmedName = node.name?.trim();
-  if (trimmedName) return trimmedName;
-  if ((node.type === "TYPE_TRIGGER" || node.type === "TYPE_ACTION") && node.component) return node.component;
-  return "node";
-}
-
-function buildDuplicatedNodes(specNodes: ComponentsNode[], nodeIds: string[]): DuplicatedNodesResult {
-  const existingNodeNames = specNodes.map((node) => node.name || "").filter(Boolean);
-  const newNodes: ComponentsNode[] = [];
-  const nodeIdMap: Record<string, string> = {};
-
-  for (const nodeId of nodeIds) {
-    const nodeToDuplicate = specNodes.find((node) => node.id === nodeId);
-    if (!nodeToDuplicate) continue;
-
-    const baseName = duplicateBaseName(nodeToDuplicate);
-    const allNames = [...existingNodeNames, ...newNodes.map((node) => node.name || "")];
-    const uniqueNodeName = generateUniqueNodeName(baseName, allNames);
-    const newNodeId = generateNodeId(baseName, uniqueNodeName);
-
-    nodeIdMap[nodeId] = newNodeId;
-    newNodes.push({
-      ...nodeToDuplicate,
-      id: newNodeId,
-      name: uniqueNodeName,
-      position: {
-        x: (nodeToDuplicate.position?.x || 0) + 50,
-        y: (nodeToDuplicate.position?.y || 0) + 50,
-      },
-      isCollapsed: false,
-    });
-  }
-
-  return { newNodes, nodeIdMap };
-}
-
-function buildDuplicatedEdges(
-  edges: ComponentsEdge[],
-  duplicatedNodeIds: Set<string>,
-  nodeIdMap: Record<string, string>,
-): ComponentsEdge[] {
-  return edges
-    .filter(
-      (edge) =>
-        edge.sourceId != null &&
-        edge.targetId != null &&
-        duplicatedNodeIds.has(edge.sourceId) &&
-        duplicatedNodeIds.has(edge.targetId),
-    )
-    .map((edge) => ({
-      ...edge,
-      sourceId: nodeIdMap[edge.sourceId!],
-      targetId: nodeIdMap[edge.targetId!],
-    }));
-}
 
 function getCanvasVersionEditPermissionState({
   canUpdateCanvasVersion,
